samples/deployment-requirements.txt
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/pre-release
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/nightly
openvino_genai~=2025.2.0.0.dev
librosa==0.10.2.post1  # For Whisper
pillow==11.1.0  # Image processing for VLMs
